fix: clean up stories files, improve docs
